@@ -1,22 +1,19 @@
 {
-  "name": "react-training-RoboDex",
+  "name": "redux-training-robodex",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
     "gh-pages": "^0.12.0",
-    "react-scripts": "0.8.4"
+    "react-scripts": "^0.9.5"
   },
   "dependencies": {
-    "fetch": "^1.1.0",
-    "immutable": "^3.8.1",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "prop-types": "^15.5.7",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-redux": "^5.0.1",
     "redux": "^3.6.0",
     "redux-logger": "^2.7.4",
-    "redux-thunk": "^2.1.0",
-    "reselect": "^2.5.4",
-    "tachyons": "^4.6.1"
+    "redux-thunk": "^2.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
