@@ -1,9 +1,10 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 const Card = ({id, name, email}) => {
   return (
     <div className='grow bg-light-green br3 pa3 ma2 dib'>
-      <img role='presentation' src={`//robohash.org/${id}?size=200x200`} />
+      <img alt={name} src={`//robohash.org/${id}?size=200x200`} />
       <div>
         <h2>{name}</h2>
         <p>{email}</p>
@@ -13,9 +14,9 @@ const Card = ({id, name, email}) => {
 };
 
 Card.propTypes = {
-  id: React.PropTypes.number.isRequired,
-  name: React.PropTypes.string.isRequired,
-  email: React.PropTypes.string.isRequired
+  id: PropTypes.number.isRequired,
+  name: PropTypes.string.isRequired,
+  email: PropTypes.string.isRequired
 
 };
 export default Card;
